+++ packages/polythene-core-card/src/card-actions.js
@@ -1,4 +1,4 @@
-import { filterSupportedAttributes } from "polythene-core";
+import { filterSupportedAttributes, deprecation } from "polythene-core";
 import classes from "polythene-css-classes/card";
 import buttonClasses from "polythene-css-classes/button";
 
@@ -6,6 +6,12 @@ const actionLayoutClasses = {
   horizontal: classes.actionsHorizontal,
   vertical:   classes.actionsVertical,
   justified:  classes.actionsJustified
+};
+
+export const onMount = ({ attrs }) => {
+  if (attrs.bordered) {
+    deprecation("Card", "bordered", "border");
+  }
 };
 
 const actionClassForLayout = (layout = "horizontal") => actionLayoutClasses[layout];
@@ -21,7 +27,7 @@ export const createProps = (vnode, { keys: k }) => {
         classes.actions,
         attrs.layout !== "vertical" ? buttonClasses.row : null,
         actionClassForLayout(attrs.layout),
-        attrs.bordered ? classes.actionsBordered : null,
+        (attrs.border || attrs.bordered) ? classes.actionsBorder : null,
         attrs.tight ? classes.actionsTight : null,
         attrs.className || attrs[k.class]
       ].join(" "),