+++ packages/polythene-core/src/deprecated.js
@@ -1,0 +1,3 @@
+
+export const deprecation = (component, deprecatedOption, newOption) =>
+  console.warn(`${component}: option '${deprecatedOption}' is deprecated and will be removed in later versions. Use '${newOption}' instead.`);