+++ packages/polythene-core/src/index.js
@@ -7,3 +7,4 @@ export { Multi } from "./multi";
 export { show, hide } from "./transition";
 export { throttle, subscribe, unsubscribe, emit } from "./events";
 export { getStyle, isRTL } from "./style";
+export { deprecation } from "./deprecated";