+++ packages/polythene-css-classes/button.js
@@ -11,7 +11,7 @@ export default {
   wash:       "pe-button__wash",
 
   // states
-  borders:    "pe-button--borders",
+  border:     "pe-button--border",
   disabled:   "pe-button--disabled",
   focused:    "pe-button--focus",
   inactive:   "pe-button--inactive",