+++ packages/polythene-css-classes/card.js
@@ -19,7 +19,7 @@ export default {
   title:               "pe-card__title",
 
   // states
-  actionsBordered:     "pe-card__actions--borders",
+  actionsBorder:       "pe-card__actions--border",
   actionsHorizontal:   "pe-card__actions--horizontal",
   actionsJustified:    "pe-card__actions--justified",
   actionsTight:        "pe-card__actions--tight",