+++ packages/polythene-css-classes/list.js
@@ -4,10 +4,10 @@ export default {
   component:       "pe-list",
 
   // states
-  borders:         "pe-list--borders",
+  border:          "pe-list--border",
   compact:         "pe-list--compact",
   hasHeader:       "pe-list--header",
-  indentedBorders: "pe-list--indented-borders",
+  indentedBorder:  "pe-list--indented-border",
   padding:         "pe-list--padding",
   paddingTop:      "pe-list--padding-top",
   paddingBottom:   "pe-list--padding-bottom",