+++ packages/polythene-css-list/src/color.js
@@ -6,7 +6,7 @@ const style = (scopes, selector, componentVars, tint) => [{
       borderTopColor: componentVars["color_" + tint + "_border"]
     },
 
-    ".pe-list--borders": {
+    ".pe-list--border": {
       " .pe-list-tile": {
         ":not(:last-child)": {
           borderColor: componentVars["color_" + tint + "_border"]
@@ -14,7 +14,7 @@ const style = (scopes, selector, componentVars, tint) => [{
       }
     },
 
-    ".pe-list--indented-borders": {
+    ".pe-list--indented-border": {
       " .pe-list-tile": {
         " .pe-list-tile__content:not(.pe-list-tile__content-front)": {
           borderColor: componentVars["color_" + tint + "_border"]